+++ CHANGELOG.rst
@@ -15,6 +15,7 @@ Alexander Schlaich, Henrik Stooß, Anirban Dutta
 
 - Added type annotations for observables in ``AnalysisBase`` and ``ProfileBase`` to
   enable mypy type checking without ``# type: ignore`` comments (#531)
+- Change DielectricCylinder code to avoid numerical errors (#556)
 - Improve guidelines for contributing (#560)
 - Removed platforms from ``tox.ini``, this is handled by the build workflow (#564)
 - Don't include the ``.github`` folder in the source distribution (#562)